auth-service/requirements.txt
boto3
bcrypt
PyJWT